move `code` to fixtures folder
