+++ test/fixtures/mocked-components/package-server-with-plugin.js
